refactor(config): split ConfigValidator.swift into per-language files
Extract 798 LOC monolith into 7 focused files:
- ValidatorDetection.swift (162 LOC) — models + detection + tool availability
- YAMLValidator.swift (150 LOC) — yamllint parser + built-in YAML checks
- ShellValidator.swift (111 LOC) — shellcheck parser + built-in shell checks
- DockerfileValidator.swift (142 LOC) — hadolint parser + built-in Dockerfile checks
- TerraformValidator.swift (60 LOC) — terraform validate parser
- ValidationWorker.swift (124 LOC) — background validation dispatch
- ConfigValidator.swift (120 LOC) — facade only

Public API unchanged. All 207 existing tests pass.
